extracted infection model to separate interface
