+++ src/main/java/org/matsim/episim/InfectionEventHandler.java
@@ -20,10 +20,7 @@ import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.gbl.Gbl;
 import org.matsim.core.router.TripStructureUtils;
 import org.matsim.episim.EpisimPerson.DiseaseStatus;
-import org.matsim.episim.model.DefaultInfectionModel;
-import org.matsim.episim.model.DefaultProgressionModel;
-import org.matsim.episim.model.InfectionModel;
-import org.matsim.episim.model.ProgressionModel;
+import org.matsim.episim.model.*;
 import org.matsim.episim.policy.ShutdownPolicy;
 import org.matsim.facilities.Facility;
 import org.matsim.utils.objectattributes.attributable.Attributes;
